# docs/swagger.yaml
definitions:
  models.APIResponse:
    properties:
      data:
        description: Optional data payload
      error:
        description: |-
          Optional error
          example: Something went wrong
        type: string
      message:
        description: |-
          The response message
          example: Success
        type: string
    type: object
...
  models.Software:
    properties:
      created_at:
        type: string
      description:
        example: Communication Suite
        type: string
      id:
        example: 1
        type: integer
      name:
        example: ClickUp
        type: string
      type:
        description: e.g., License, SaaS, etc.
        example: SaaS
        type: string
      updated_at:
        type: string
    type: object
...
  models.SoftwareAssignmentLog:
    properties:
      action:
        description: Assigned | Unassigned
        example: Assigned
        type: string
      changed_at:
        example: "2025-06-11T15:04:05Z"
        type: string
      changed_by:
        example: 2
        type: integer
      id:
        example: 1
        type: integer
      software:
        $ref: '#/definitions/models.Software'
      software_id:
        example: 7
        type: integer
      staff:
        $ref: '#/definitions/models.StaffPlain'
      staff_id:
        example: 3
        type: integer
      updated_at:
        example: "2025-06-11T15:05:00Z"
        type: string
    type: object
...
  models.StaffPlain:
    properties:
      created_at:
        type: string
      department_id:
        example: 2
        type: integer
      email:
        example: john.doe@shuttlers.co
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 1
        type: integer
      last_name:
        example: Doe
        type: string
      status:
        example: Active
        type: string
      team_id:
        example: 5
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/logs:
    get:
      description: Retrieves all software assignment logs with associated staff and
        software
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SoftwareAssignmentLog'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get all software assignment logs
      tags:
      - Software Assignment Logs
    post:
      consumes:
      - application/json
      description: Adds a new log entry for a software assignment action
      parameters:
      - description: Software Assignment Log
        in: body
        name: log
        required: true
        schema:
          $ref: '#/definitions/models.SoftwareAssignmentLog'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SoftwareAssignmentLog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Create a software assignment log
      tags:
      - Software Assignment Logs
  /api/logs/{id}:
    delete:
      description: Deletes a log entry by ID
      parameters:
      - description: Log ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Delete a software assignment log
      tags:
      - Software Assignment Logs
    get:
      description: Retrieves a single software assignment log by its ID
      parameters:
      - description: Log ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SoftwareAssignmentLog'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get a software assignment log by ID
      tags:
      - Software Assignment Logs
    put:
      consumes:
      - application/json
      description: Updates an existing log entry by ID
      parameters:
      - description: Log ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Log
        in: body
        name: log
        required: true
        schema:
          $ref: '#/definitions/models.SoftwareAssignmentLog'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SoftwareAssignmentLog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Update a software assignment log
      tags:
      - Software Assignment Logs
  /api/logs/details:
    get:
      description: Retrieves logs with optional filtering by date and search query
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end
        type: string
      - description: Search by staff or software name
        in: query
        name: search
        type: string
      - description: Limit results
        in: query
        name: limit
        type: integer
      - description: Offset results
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SoftwareAssignmentLog'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get filtered and paginated logs
      tags:
      - Software Assignment Logs
  /api/logs/software/{id}:
    get:
      description: Paginated and filtered logs for a specific software ID
      parameters:
      - description: Software ID
        in: path
        name: id
        required: true
        type: integer
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Search by staff name
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SoftwareAssignmentLog'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get detailed logs for software
      tags:
      - Software Assignment Logs
  /api/logs/software/{id}/details/raw:
    get:
      description: Raw SQL version of software logs query
      parameters:
      - description: Software ID
        in: path
        name: id
        required: true
        type: integer
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Search by staff name
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get raw logs for software
      tags:
      - Software Assignment Logs
  /api/logs/software/{id}/plain:
    get:
      description: Retrieves all logs for a specific software ID
      parameters:
      - description: Software ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SoftwareAssignmentLog'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get logs for specific software
      tags:
      - Software Assignment Logs
  /api/logs/staff/{id}:
    get:
      description: Retrieves all logs for a specific staff ID
      parameters:
      - description: Staff ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SoftwareAssignmentLog'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get logs for specific staff
      tags:
      - Software Assignment Logs
  /api/logs/staff/{id}/details:
    get:
      description: Paginated and filtered logs for a given staff ID
      parameters:
      - description: Staff ID
        in: path
        name: id
        required: true
        type: integer
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Search by software name
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SoftwareAssignmentLog'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get detailed logs for staff
      tags:
      - Software Assignment Logs
  /api/logs/staff/{id}/details/raw:
    get:
      description: Raw SQL version of staff logs query
      parameters:
      - description: Staff ID
        in: path
        name: id
        required: true
        type: integer
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Search by software name
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get raw logs for staff
      tags:
      - Software Assignment Logs
...
swagger: "2.0"